# apiv1/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Momingo Api
# during dev, should point to your local machine
#host: aldair-218814.appspot.com
host: localhost:10010
#GRPCS
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - multipart/form-data
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /rateposts:
    x-swagger-router-controller: rateposts.controller
    post:
      description: Agregar una calificación o modifica a un post
      operationId: ratePost
      parameters:
        - in: body
          name: Post Ratting body
          required: true
          schema:
            $ref: "#/definitions/PostRate"
      responses:
        "200":
          description: Ha sido completado
          schema:
            $ref: "#/definitions/GeneralResponse"
        "409":
          description: No se ha completado
          schema:
            $ref: "#/definitions/GeneralResponse"
  /saveposts:
    x-swagger-router-controller: saveposts.controller
    post:
      description: Guarda un post para poder verlo después (No en su dispostivo sólo en su cuenta).
      operationId: savePost
      parameters:
        - in: body
          name: PostSave body
          required: true
          schema:
            $ref: "#/definitions/PostSave"
      responses:
        "200":
          description: Ha sido completado
          schema:
            $ref: "#/definitions/GeneralResponse"
        "409":
          description: No se ha completado
          schema:
            $ref: "#/definitions/GeneralResponse"
    delete:
      description: Borra de sus guardados alguna publicación
      operationId: deletesavedPost
      parameters:
        - name: id_save
          in: query
          type: number
          required: true
      responses:
        "200":
          description: Ha sido completado
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: No se ha completado
          schema:
            $ref: "#/definitions/GeneralResponse"
  /posts:
    # binds a127 app logic to a route
    x-swagger-router-controller: posts.controller
    get:
      description: Retorna la lista de publicaciones de un usuario
      operationId: getPosts
      parameters:
        - name: username
          type: string
          required: true
          in: query
      responses:
        "200":
          description: Los post de un usuario
          schema:
            $ref: "#/definitions/PostList"
        "409":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Borra una publicación
      operationId: deletePost
      parameters:
        - name: id_post
          type: number
          required: true
          in: query
      responses:
        "200":
          description: Se ha borrado
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: No se pudo, man
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: createPost
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: file
          in: formData
          description: Meme que se va a subir
          type: file
        - name: body
          in: formData
          description: Cuerpo del post, username y descripcion
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /login:
    x-swagger-router-controller: user.controller
    get:
      description: Loggeo de un usuario
      operationId: getUserLog
      parameters:
        - name: username
          in: query
          type: string
          required: true
        - name: pass
          in: query
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/User"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /usuarios:
    x-swagger-router-controller: user.controller
    post:
      description: Registro de un usuario a la BD
      operationId: createUser
      parameters:
        - in: body
          name: Usuario
          schema:
           $ref: "#/definitions/User"
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/User'
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      description: Retorna lista de usuarios
      operationId: getUsers
      parameters:
        - in: query
          name: data
          type: string
          required: false
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUsersList"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: delete a user
      operationId: deleteUser
      parameters:
        - name: username
          description: username
          type: string
          in: query
          required: true
        - name: password
          description: password
          type: string
          in: query
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/User"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /usuarios/{id}:
    x-swagger-router-controller: user.controller
    put:
      description: update an user data
      operationId: updateUser
      parameters:
        - name: id
          description: El nombre del usuario ejemplo "aldairmh"
          type: string
          in: path
          required: true
        - in: body
          name: usuario
          description: The Operador to update.
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/User"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  

  /follows:
    x-swagger-router-controller: follows.controller
    get:
      description: Ver los seguidores de un usuario
      operationId: listFollowers
      parameters:
        - name: username
          in: query
          type: string
          required: true
      responses:
        "200":
          description: List of followers
          schema:
            $ref: "#/definitions/ListFollowers"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Dejar de seguir a un usuario
      operationId: deleteFollow
      parameters:
        - in: query
          name: username_seguidor
          type: string
          required: true
        - in: query
          name: username_seguido
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    post:
      description: Seguir a otro usuario
      operationId: seguirUsuario
      parameters:
        - in: body
          name: follow
          schema:
            $ref: "#/definitions/Follow"
      responses:
        "200":
          description: SUCCESS
          schema:
            $ref: "#/definitions/GeneralResponse"
        "409":
          description: SQLException
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: ERROR
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  PostList:
    required:
      - posts
    properties:
      posts:
        type: array
        items:
          $ref: "#/definitions/User"

  PostRate:
    required:
      - username
      - post
      - calificacion
    properties:
      username:
        type: string
      post:
        type: number
      calificacion:
        type: number
  PostSave:
    required:
      - username
      - id
    properties:
      username:
        type: string
      id:
        type: number
  Follow:
    required:
      - followerid
      - followedid
    properties:
      followerid:
        type: string
      followedid:
        type: string
  Follow2:
    properties:
      username_seguidor:
        type: string
      username_seguido:
        type: string
  ListFollowers:
    required:
      - followers
    properties:
      followers:
        type: array
        items:
          $ref: "#/definitions/Follow2"
  Post:
    required:
      - username
    properties:
      username:
        type: string
      descripcion:
        type: string
  GetUsersList:
    required:
      - users
    properties:
      users:
        type: array
        items:
          $ref: "#/definitions/User"
  UserProfile:
    type: object
    properties:
      username:
        type: string
      email:
        type: string
      descripcion:
        type: string
  User:
    type: object
    properties:
      username:
        type: string
      email:
        type: string
      descripcion:
        type: string
      passw:
        type: string
  GeneralResponse:
    properties:
      message:
        type: string
  ErrorIntResponse:
    properties:
      code:
        type: number
      message:
        type: string
      description:
        type: string
    required:
      - message
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string